changed float to float32

+++ rktl_sim/src/simulator/car.py
@@ -105,7 +105,7 @@ class Car(object):
         heading = p.getEulerFromQuaternion(orient)[2]
         r_inv = np.array([[math.cos(heading), -math.sin(heading), 0.],
                           [math.sin(heading), math.cos(heading), 0.],
-                         [0., 0., 1.]], dtype=np.float)
+                          [0., 0., 1.]], dtype=np.float32)
         linear = r_inv @ linear
         return (linear, angular)
